requirements.txt
openapi-python-client
discord
discord.py
python-dotenv
beautifulsoup4
aiohttp
